packages/modeldb: add subscribe and unsubscribe
