+++ packages/modeldb/package.json
@@ -27,10 +27,8 @@
   "dependencies": {
     "@canvas-js/signed-cid": "0.5.0-alpha4",
     "@libp2p/logger": "^2.1.1",
-    "@noble/hashes": "^1.3.0",
     "better-sqlite3": "^8.5.2",
     "idb": "^7.1.1",
-    "multiformats": "^12.1.0",
-    "uint8arrays": "^4.0.6"
+    "multiformats": "^12.1.0"
   }
 }
