+++ packages/modeldb/src/browser/utils.ts
@@ -1,0 +1,3 @@
+export const getObjectStoreName = (model: string) => `record/${model}`
+export const getTombstoneObjectStoreName = (model: string) => `tombstone/${model}`
+export const getIndexName = (index: string[]) => index.join("/")
